+++ packages/react-ui/src/widgets/CategoryWidgetUI/CategoryWidgetUI.styled.js
@@ -1,11 +1,12 @@
-import { Box, Grid, Link, styled } from '@mui/material';
+import { Box, Button, Grid, Link, styled } from '@mui/material';
 import Typography from '../../components/atoms/Typography';
 
 const REST_CATEGORY = '__rest__';
 
-export const CategoriesWrapper = styled(Grid)(({ theme: { spacing } }) => ({
-  maxHeight: spacing(40),
-  padding: spacing(0, 1, 1, 0)
+export const CategoriesWrapper = styled(Grid)(({ theme }) => ({
+  maxHeight: theme.spacing(40),
+  overflow: 'auto',
+  padding: 0
 }));
 
 export const CategoryItemGroup = styled(Grid, {
@@ -13,6 +14,19 @@ export const CategoryItemGroup = styled(Grid, {
 })(({ theme, selectable, name, unselected }) => {
   return {
     flexDirection: 'row',
+    maxWidth: '100%',
+    padding: theme.spacing(0.5, 0.25),
+    margin: 0,
+
+    '> .MuiGrid-item': {
+      paddingTop: 0,
+      paddingLeft: 0
+    },
+    '&:focus-visible': {
+      outline: `none !important`,
+      boxShadow: `inset 0 0 0 2px ${theme.palette.primary.main} !important`
+    },
+
     ...(unselected && {
       color: theme.palette.text.disabled,
 
@@ -54,7 +68,7 @@ export const OptionsSelectedBar = styled(Grid)(({ theme: { spacing, palette } })
 export const ProgressBar = styled(Grid)(({ theme }) => ({
   height: theme.spacing(0.5),
   width: '100%',
-  margin: theme.spacing(0.5, 0, 1, 0),
+  margin: theme.spacing(0.5, 0, 0.25, 0),
   borderRadius: theme.spacing(0.5),
   backgroundColor: theme.palette.action.disabledBackground,
 
@@ -67,6 +81,16 @@ export const ProgressBar = styled(Grid)(({ theme }) => ({
                  width ${theme.transitions.easing.sharp} ${theme.transitions.duration.complex}ms`
   }
 }));
+
+export const CategoryLabelWrapper = styled(Grid, {
+  shouldForwardProp: (prop) => prop !== 'isSelectable'
+})(({ theme, isSelectable }) => {
+  return {
+    ...(isSelectable && {
+      width: `calc(100% - ${theme.spacing(4)})`
+    })
+  };
+});
 
 export const CategoryLabel = styled(Typography)(({ theme }) => ({
   fontWeight: theme.typography.fontWeightBold,
@@ -84,4 +108,20 @@ export const LinkAsButton = styled(Link)(({ theme }) => ({
 
 export const CategoriesRoot = styled(Box)(({ theme }) => ({
   ...theme.typography.body2
+}));
+
+export const HiddenButton = styled(Button)(({ theme }) => ({
+  position: 'absolute',
+  left: '-999px',
+  top: '-1px',
+  width: '1px',
+  height: '1px',
+  display: 'inline-flex',
+
+  '&:focus-visible': {
+    position: 'static',
+    width: 'auto',
+    height: 'auto',
+    marginTop: theme.spacing(2)
+  }
 }));